Non-functional merge of Extractor files in expression module
